requirements-file.txt
# requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.9
cryptography==41.0.1
pillow==9.5.0
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.24.3
opencv-python==4.7.0.72
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
